Domain entities updated

+++ Domain/Scheduler/Worker.cs
@@ -2,11 +2,11 @@ using System;
 
 namespace Domain.Scheduler
 {
-    public record Worker(Guid Id, string Name, string PathToIcon, double EfficiencyCoeff)
+    public record Worker(Guid Id, string Name, string PathToIcon, float EfficiencyCoeff)
     {
         public Guid Id { get; } = Id;
         public string Name { get; } = Name;
         public string PathToIcon { get; } = PathToIcon;
-        public double EfficiencyCoeff { get; } = EfficiencyCoeff;
+        public float EfficiencyCoeff { get; } = EfficiencyCoeff;
     }
 }
